Separate custom array functions
